Version 0.8.8
This is the version 0.8.8 for steam as of 2018/02/17. Below is the announcement from Eek! Games.

0.8.8 Redeploy
Hey guys,

We made the decision to include some last minute updates we've been working on to 0.8.8 since they ended up being ready to go. We are not incrementing the game version so as to preserve save files. We have greatly improved a lot of performance aspects of the latest updates we've put out, and for those of you maxing out on FPS or getting input lag, you should now have a much smoother experience. 

CHANGE LOG:
Fix for Vickie running away during Never Have I Ever
Fix for gutter not being clickable
Optimizations and performance tweaks for new content
Dialogue audio now streams from HDD (smaller memory footprint)
Added ambient yard noises